refactoring pods and nodes functions
